separate routesInfo from app router
